requirements.txt
google-generativeai>=0.4.0
docker>=7.0.0
requests>=2.31.0
python-dotenv>=1.0.0